third_party/blink/web_tests/external/wpt/css/css-page/page-rule-declarations-004-expected.txt

This is a testharness.js-based test.
[FAIL] rule with invalid selector ['--a']
  assert_not_equals: no @page rule should have been parsed got disallowed value 6
[FAIL] assigning invalid selector text ['--a']
  assert_equals: should not be able to assign an invalid selector expected "" but got "--a"
Harness: the test ran to completion.
